Added a move function.
This doesn't do anything as of yet, since it is(will) be dependent on the PickPawn function where you pick what pawn you want to move. As of yet it is not fully implemented, jsut some early test the MovePawn function doesn't do much, at least it is not crashing the program. ¯\_(ツ)_/¯