Fixed the movePawn function
Fixed an error causing and infinite loop in the movePawn function. This was caused by some liens of code I've forgotten to write that compared if(amongst other) pawnLoc[5] was les than 21 or not. Added an alert making it more clear to the uset that the pawn is in goal if the number is more than 20.