AWESOME MovePawn update
Instead of defining what pawn to move by a defined number. Instead of making if(pawnLoc[0]) for example, I've just changed it to if(pawnLoc[(currentPawn -1)] so that I can add theoretically any number of paws to the board without changing this part of the code. This will future proof if I decide to add more pawns. I was able to do it since I split up each players pawnLoc array into pawnLoc1 and pawnLoc2. In the old function I had to add an else if each time, adding up to about 3 lines of code.