Experimenting in ways to se if move is valid
This is the point where I'd drink coffee if I were a coffe drinker.. have experimented in ways to get the program to validate a move. but the validation process is messy and get looong lines of code. Instead I'll opt for a function within the movePawn function to sort it.