Validation of players move and some fixes for it
changed the prompt asking for what pawn to move to have parsInt making int an int instead of a string. Seemed to solve some calculation issues (?) in the movePawn function. Also changed the move cariable to be global so that I coud use it in other functions, for now it is used in the validateMove function.