Bug fix in the validateMove().
the function checks if a pawn is already occupying a square by a pawn of the same player. The function is however checking if the pawnlocation(aka, how many steps the pawn has taken) is the same even if it has left the board. which is irrelevant. So, added an additional condition to be true(condition beeing that the targeted location is <=15, in other words outside the board) to invalidate the board: